refactor: テストを src/ から tests/ ディレクトリに移行し、Clippy を厳格化
- 全クレートのルートに #![deny(warnings)] を追加
- #[cfg(test)] mod tests ブロックを src/ から tests/ に移行
- テストのインポートパスを公開 API 経由に修正
- プライベート関数のテストを公開 API テストに置き換え
- drawio_renderer の extract_style_value を pub(crate) に変更
- i18n の doc test を ignore マーク
