+++ crates/katana-core/src/lib.rs
@@ -1,3 +1,5 @@
+#![deny(warnings)]
+
 pub mod ai;
 pub mod document;
 pub mod markdown;
