+++ crates/katana-core/src/markdown/drawio_renderer.rs
@@ -216,8 +216,18 @@ fn render_edge(
         return;
     };
 
-    let src = Rect { x: sx, y: sy, w: sw, h: sh };
-    let tgt = Rect { x: tx, y: ty, w: tw, h: th };
+    let src = Rect {
+        x: sx,
+        y: sy,
+        w: sw,
+        h: sh,
+    };
+    let tgt = Rect {
+        x: tx,
+        y: ty,
+        w: tw,
+        h: th,
+    };
     let (scx, scy) = src.center();
     let (tcx, tcy) = tgt.center();
     let waypoints = collect_waypoints(cell);
@@ -346,73 +356,4 @@ fn xml_escape(s: &str) -> String {
     s.replace('&', "&amp;")
         .replace('<', "&lt;")
         .replace('>', "&gt;")
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use crate::markdown::diagram::{DiagramBlock, DiagramKind};
-
-    const SIMPLE_DRAWIO_XML: &str = r#"<mxfile><diagram name="test"><mxGraphModel><root>
-<mxCell id="0"/>
-<mxCell id="1" parent="0"/>
-<mxCell id="2" value="Box A" style="rounded=1;fillColor=#fff2cc;strokeColor=#d6b656;" vertex="1" parent="1">
-    <mxGeometry x="80" y="80" width="120" height="60" as="geometry"/>
-</mxCell>
-<mxCell id="3" value="Box B" vertex="1" parent="1">
-    <mxGeometry x="280" y="80" width="120" height="60" as="geometry"/>
-</mxCell>
-</root></mxGraphModel></diagram></mxfile>"#;
-
-    #[test]
-    fn 有効なdrawio_xmlがsvgに変換される() {
-        let block = DiagramBlock {
-            kind: DiagramKind::DrawIo,
-            source: SIMPLE_DRAWIO_XML.to_string(),
-        };
-        let result = render_drawio(&block);
-        assert!(matches!(result, DiagramResult::Ok(_)));
-        if let DiagramResult::Ok(html) = result {
-            assert!(html.contains("<svg"));
-            assert!(html.contains("Box A"));
-            assert!(html.contains("Box B"));
-        }
-    }
-
-    #[test]
-    fn 無効なxmlはエラー結果を返す() {
-        let block = DiagramBlock {
-            kind: DiagramKind::DrawIo,
-            source: "not xml".to_string(),
-        };
-        let result = render_drawio(&block);
-        assert!(matches!(result, DiagramResult::Err { .. }));
-    }
-
-    #[test]
-    fn 楕円スタイルが処理される() {
-        let xml = r#"<mxGraphModel><root>
-<mxCell id="0"/><mxCell id="1" parent="0"/>
-<mxCell id="2" value="Circle" style="ellipse;" vertex="1" parent="1">
-    <mxGeometry x="50" y="50" width="100" height="100" as="geometry"/>
-</mxCell>
-</root></mxGraphModel>"#;
-        let block = DiagramBlock {
-            kind: DiagramKind::DrawIo,
-            source: xml.to_string(),
-        };
-        let result = render_drawio(&block);
-        if let DiagramResult::Ok(html) = result {
-            assert!(html.contains("<ellipse"));
-        }
-    }
-
-    #[test]
-    fn spaceの抽出() {
-        assert_eq!(
-            extract_style_value("rounded=1;fillColor=#fff;", "fillColor"),
-            Some("#fff")
-        );
-        assert_eq!(extract_style_value("rounded=1;", "fillColor"), None);
-    }
 }
