+++ crates/katana-core/src/markdown/mermaid_renderer.rs
@@ -118,41 +118,3 @@ fn create_input_file(source: &str) -> Result<NamedTempFile, String> {
         .map_err(|e| format!("一時ファイル書き込み失敗: {e}"))?;
     Ok(file)
 }
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use crate::markdown::diagram::{DiagramBlock, DiagramKind};
-
-    #[test]
-    fn mmdc未検出時はエラー結果を返す() {
-        // 存在しないバイナリを指定してフォールバックを検証する。
-        std::env::set_var("MERMAID_MMDC", "/nonexistent/mmdc");
-        let block = DiagramBlock {
-            kind: DiagramKind::Mermaid,
-            source: "graph TD; A-->B".to_string(),
-        };
-        let result = render_mermaid(&block);
-        assert!(matches!(result, DiagramResult::CommandNotFound { .. }));
-        // テスト後に環境変数を戻す。
-        std::env::remove_var("MERMAID_MMDC");
-    }
-
-    // mmdc がシステムで利用可能な場合のみ実行する結合テスト。
-    #[test]
-    fn mmdcが利用可能なら正しくsvgを返す() {
-        // MERMAID_MMDC が nonexistent になっている場合はスキップ。
-        if std::env::var("MERMAID_MMDC").as_deref() == Ok("/nonexistent/mmdc") {
-            return;
-        }
-        if !is_mmdc_available() {
-            return; // mmdc が未インストールならスキップ。
-        }
-        let block = DiagramBlock {
-            kind: DiagramKind::Mermaid,
-            source: "graph TD; A-->B".to_string(),
-        };
-        let result = render_mermaid(&block);
-        assert!(matches!(result, DiagramResult::OkPng(_)));
-    }
-}
