+++ crates/katana-ui/src/preview_pane.rs
@@ -210,25 +210,22 @@ impl PreviewPane {
 
     /// テスト用: Pending がなくなるまでバックグラウンドスレッドをブロック待機する。
     #[cfg(test)]
+    #[allow(dead_code)]
     pub fn wait_for_renders(&mut self) {
-        loop {
-            if let Some(rx) = &self.render_rx {
-                while let Ok((idx, section)) = rx.try_recv() {
-                    if idx < self.sections.len() {
-                        self.sections[idx] = section;
-                    }
+        while let Some(rx) = &self.render_rx {
+            while let Ok((idx, section)) = rx.try_recv() {
+                if idx < self.sections.len() {
+                    self.sections[idx] = section;
                 }
-                if self
-                    .sections
-                    .iter()
-                    .any(|s| matches!(s, RenderedSection::Pending { .. }))
-                {
-                    std::thread::sleep(std::time::Duration::from_millis(50));
-                } else {
-                    self.render_rx = None;
-                    break;
-                }
+            }
+            if self
+                .sections
+                .iter()
+                .any(|s| matches!(s, RenderedSection::Pending { .. }))
+            {
+                std::thread::sleep(std::time::Duration::from_millis(50));
             } else {
+                self.render_rx = None;
                 break;
             }
         }
@@ -410,7 +407,7 @@ fn try_rasterize(kind: &DiagramKind, source: &str, html: &str) -> RenderedSectio
 }
 
 /// HTML フラグメントから `<svg...>...</svg>` を抽出する。
-fn extract_svg(html: &str) -> Option<&str> {
+pub fn extract_svg(html: &str) -> Option<&str> {
     let start = html.find("<svg")?;
     let end = html.rfind("</svg>")? + "</svg>".len();
     Some(&html[start..end])
@@ -435,7 +432,7 @@ fn decode_png_to_section(kind: &DiagramKind, source: &str, bytes: Vec<u8>) -> Re
 }
 
 /// PNG バイト列を RGBA ピクセルに変換する。
-fn decode_png_rgba(bytes: &[u8]) -> Result<RasterizedSvg, String> {
+pub fn decode_png_rgba(bytes: &[u8]) -> Result<RasterizedSvg, String> {
     let img = image::load_from_memory(bytes).map_err(|e| e.to_string())?;
     let rgba = img.into_rgba8();
     let (width, height) = rgba.dimensions();
@@ -444,160 +441,4 @@ fn decode_png_rgba(bytes: &[u8]) -> Result<RasterizedSvg, String> {
         height,
         rgba: rgba.into_raw(),
     })
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    /// ヘルパー: RenderedSection から Markdown テキストを抽出する。
-    fn markdown_texts(sections: &[RenderedSection]) -> Vec<&str> {
-        sections
-            .iter()
-            .filter_map(|s| match s {
-                RenderedSection::Markdown(md) => Some(md.as_str()),
-                _ => None,
-            })
-            .collect()
-    }
-
-    // ── 3.2 プレビュー同期: 未保存バッファからの即時プレビュー更新 ──
-
-    #[test]
-    fn 未保存バッファの変更がプレビューに反映される() {
-        let mut pane = PreviewPane::default();
-
-        // 初期コンテンツでプレビューを構築
-        pane.update_markdown_sections("# Hello");
-        assert_eq!(pane.sections.len(), 1);
-        let texts = markdown_texts(&pane.sections);
-        assert!(texts[0].contains("# Hello"));
-
-        // ファイル保存なしでバッファを更新 → プレビューに反映される
-        pane.update_markdown_sections("# Hello World\n\nNew paragraph");
-        let texts = markdown_texts(&pane.sections);
-        assert!(texts[0].contains("# Hello World"));
-        assert!(texts[0].contains("New paragraph"));
-    }
-
-    #[test]
-    fn 連続編集がプレビューに即時反映される() {
-        let mut pane = PreviewPane::default();
-
-        // 複数回の連続編集がすべて反映される
-        let edits = vec![
-            "# Draft 1",
-            "# Draft 2\n\n- item A",
-            "# Draft 3\n\n- item A\n- item B\n- item C",
-        ];
-
-        for edit in &edits {
-            pane.update_markdown_sections(edit);
-            let texts = markdown_texts(&pane.sections);
-            assert!(
-                texts[0].contains(edit),
-                "Edit not reflected in preview: {edit}"
-            );
-        }
-    }
-
-    #[test]
-    fn 空バッファでもプレビューがクラッシュしない() {
-        let mut pane = PreviewPane::default();
-
-        // コンテンツを入力
-        pane.update_markdown_sections("# Hello");
-        assert_eq!(pane.sections.len(), 1);
-
-        // 空に戻す → セクション数は 0 になる（空文字列は flush されない）
-        pane.update_markdown_sections("");
-        assert_eq!(pane.sections.len(), 0);
-    }
-
-    #[test]
-    fn ダイアグラム含みのバッファでmarkdown部分のみ即時更新される() {
-        let mut pane = PreviewPane::default();
-
-        // ダイアグラムを含む初期コンテンツ
-        let source = "# Title\n```mermaid\ngraph TD; A-->B\n```\n## Footer";
-        pane.full_render(source);
-
-        // ダイアグラムは Pending になっている
-        assert!(pane.sections.len() >= 3);
-        assert!(matches!(pane.sections[1], RenderedSection::Pending { .. }));
-
-        // Markdown 部分のみ更新する（ダイアグラムは保持）
-        let modified = "# Updated Title\n```mermaid\ngraph TD; A-->B\n```\n## Updated Footer";
-        pane.update_markdown_sections(modified);
-
-        // Markdown テキストが更新されていることを確認
-        let texts = markdown_texts(&pane.sections);
-        assert!(texts.iter().any(|t| t.contains("Updated Title")));
-        assert!(texts.iter().any(|t| t.contains("Updated Footer")));
-    }
-
-    #[test]
-    fn full_renderでセクションが正しく分割される() {
-        let mut pane = PreviewPane::default();
-
-        let source = "Before\n```mermaid\ngraph TD; A-->B\n```\nAfter";
-        pane.full_render(source);
-
-        // 3 セクション: Markdown, Diagram(Pending), Markdown
-        assert_eq!(pane.sections.len(), 3);
-        assert!(matches!(pane.sections[0], RenderedSection::Markdown(_)));
-        assert!(matches!(pane.sections[1], RenderedSection::Pending { .. }));
-        assert!(matches!(pane.sections[2], RenderedSection::Markdown(_)));
-    }
-
-    #[test]
-    fn ダイアグラムなしのバッファではpendingセクションが生成されない() {
-        let mut pane = PreviewPane::default();
-
-        pane.full_render("# Pure Markdown\n\nNo diagrams here.");
-
-        assert!(pane
-            .sections
-            .iter()
-            .all(|s| matches!(s, RenderedSection::Markdown(_))));
-        assert!(!pane
-            .sections
-            .iter()
-            .any(|s| matches!(s, RenderedSection::Pending { .. })));
-    }
-
-    #[test]
-    fn プレビュー更新がファイル保存に依存しないことの検証() {
-        // Document + PreviewPane の統合テスト:
-        // ドキュメントのバッファを更新（is_dirty = true）し、save を呼ばずに
-        // プレビューが最新バッファを反映していることを確認する。
-        use katana_core::document::Document;
-
-        let mut doc = Document::new("/workspace/spec.md", "# Original");
-        let mut pane = PreviewPane::default();
-
-        // 初期プレビュー
-        pane.update_markdown_sections(&doc.buffer);
-        let texts = markdown_texts(&pane.sections);
-        assert!(texts[0].contains("# Original"));
-
-        // ドキュメントを編集（未保存状態）
-        doc.update_buffer("# Modified by user\n\nThis is not saved yet.");
-        assert!(doc.is_dirty, "ドキュメントは dirty でなければならない");
-
-        // 未保存バッファでプレビューを更新
-        pane.update_markdown_sections(&doc.buffer);
-        let texts = markdown_texts(&pane.sections);
-        assert!(
-            texts[0].contains("Modified by user"),
-            "未保存の編集がプレビューに反映されていない"
-        );
-        assert!(
-            texts[0].contains("This is not saved yet"),
-            "未保存の編集がプレビューに反映されていない"
-        );
-
-        // ドキュメントはまだ dirty のまま（保存していない）
-        assert!(doc.is_dirty, "ドキュメントは保存されていないはず");
-    }
 }
